Guard reading rcsThrustAvailable with a null-check
With the removal of the code setting / updating rcsThrustAvailable, a
bug was introduced because rcsThrustAvailable appears to always be null.

This manifests when, on my install, I:
1. perform a GravityTurn launch, not touching any throttle myself at all
2. reach a trajectory while in-atmo with an Ap. of 100km
3. wait until the vessel is between 70km and 100km (out of atmo)
4. stage into a stage with only RCS, no normal engines
5. enable RCS

Without knowing more about the internals of this mod, I do not dare to
touch it beyond putting a defensive if-statement around the update of
thrustVectorMaxThrottle.

The boolean that this method sets does not appear to be ever used inside
the mod, but it's declared public so I didn't dare to comment it out
entirely.